Move logic from Function into controller
We want the Function directories to only be concerned with Azure, not
application logic/semantics. So we moved validation of inputs into the
controller.

Jira ticket: CAMS-47
