+++ backend/functions/case-assignments/case.assignment.function.ts
@@ -3,14 +3,9 @@ import { CaseAssignmentController } from '../lib/adapters/controllers/case.assig
 import { httpError, httpSuccess } from '../lib/adapters/utils/http';
 import { applicationContextCreator } from '../lib/adapters/utils/application-context-creator';
 import log from '../lib/adapters/services/logger.service';
-import { CaseAssignmentRole } from '../lib/adapters/types/case.assignment.role';
 import { AssignmentException } from '../lib/use-cases/assignment.exception';
 
 const NAMESPACE = 'CASE-ASSIGNMENT-FUNCTION' as const;
-const INVALID_ROLE_MESSAGE =
-  'Invalid role for the attorney. Requires role to be a TrialAttorney for case assignment. ';
-const VALID_CASEID_PATTERN = RegExp('^\\d{2}-\\d{5}$');
-const INVALID_CASEID_MESSAGE = 'caseId must be formatted like 01-12345. ';
 
 const httpTrigger: AzureFunction = async function (
   functionContext: Context,
@@ -21,7 +16,6 @@ const httpTrigger: AzureFunction = async function (
   const role = request.body && request.body.role;
 
   try {
-    validateRequestParameters(caseId, functionContext, listOfAttorneyNames, role);
     await handlePostMethod(functionContext, caseId, listOfAttorneyNames, role);
   } catch (e) {
     if (e instanceof AssignmentException) {
@@ -33,42 +27,12 @@ const httpTrigger: AzureFunction = async function (
   }
 };
 
-function validateRequestParameters(
-  caseId: string,
+async function handlePostMethod(
   functionContext: Context,
+  caseId: string,
   listOfAttorneyNames: string[],
-  role: string,
+  role,
 ) {
-  const badParams = [];
-  let errors = false;
-  let message = '';
-  if (!caseId) {
-    badParams.push('caseId');
-    errors = true;
-  } else if (!caseId.match(VALID_CASEID_PATTERN)) {
-    message += INVALID_CASEID_MESSAGE;
-    errors = true;
-  }
-  if (!listOfAttorneyNames || listOfAttorneyNames.length < 1) {
-    badParams.push('attorneyList');
-    errors = true;
-  }
-  if (!role) {
-    badParams.push('role');
-    errors = true;
-  }
-  if (!(role in CaseAssignmentRole)) {
-    message += INVALID_ROLE_MESSAGE;
-    errors = true;
-  }
-  if (errors) {
-    if (badParams.length > 0)
-      message += `Required parameter(s) ${badParams.join(', ')} is/are absent.`;
-    throw new AssignmentException(400, message.trim());
-  }
-}
-
-async function handlePostMethod(functionContext: Context, caseId, listOfAttorneyNames, role) {
   const caseAssignmentController: CaseAssignmentController = new CaseAssignmentController(
     functionContext,
   );
